addressing Review comments
